Introduce LaCasa plugin
The project already mostly conformed to the rather lax requirements that
the LaCasa plugin places on unrelated code. In this instance, there were
multiple places where Throwable was used as a catch-all, causing
the risk of LaCasa's subclass of ControlThrowable to be caught.

By adding Box.uncheckedCatchControl, it's possible to declare to the
plugin that there's nothing wrong with that particular catch. It can
also be used to opt out of safety.